[hwtracing] intel_th: msu: Introduce buffer interface
Message-id: <20191118090749.8127-35-jolsa@redhat.com>
Patchwork-id: 287586
O-Subject: [PATCH RHEL8.2 BZ 1771015 34/46] intel_th: msu: Introduce buffer interface
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit 615c164da0eb42cbfb1688cb429cc4d5039db5d8
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri Jul 5 17:14:21 2019 +0300

description
===========
Introduces a concept of external buffers, which is a mechanism for creating
trace sinks that would receive trace data from MSC buffers and transfer it
elsewhere.

A external buffer can implement its own window allocation/deallocation if
it has to. It must provide a callback that's used to notify it when a
window fills up, so that it can then start a DMA transaction from that
window 'elsewhere'. This window remains in a 'locked' state and won't be
used for storing new trace data until the buffer 'unlocks' it with a
provided API call, at which point the window can be used again for storing
trace data.

This relies on a functional "last block" interrupt, so not all versions of
Trace Hub can use this feature, which does not reflect on existing users.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>