[x86] perf/x86/intel: Support PEBS output to PT
Message-id: <20191126133941.32492-54-mpetlan@redhat.com>
Patchwork-id: 289083
O-Subject: [RHEL8.2 BZ1743210 v2 053/327] perf/x86/intel: Support PEBS output to PT
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 42880f726c66f13ae1d9ac9ce4c43abe64ecac84
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Tue Aug 6 11:46:01 2019 +0300

description
===========
If PEBS declares ability to output its data to Intel PT stream, use the
aux_output attribute bit to enable PEBS data output to PT. This requires
a PT event to be present and scheduled in the same context. Unlike the
DS area, the kernel does not extract PEBS records from the PT stream to
generate corresponding records in the perf stream, because that would
require real time in-kernel PT decoding, which is not feasible. The PMI,
however, can still be used.

The output setting is per-CPU, so all PEBS events must be either writing
to PT or to the DS area, therefore, in case of conflict, the conflicting
event will fail to schedule, allowing the rotation logic to alternate
between the PEBS->PT and PEBS->DS events.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>