[fs] sysfs: Fix internal_create_group() for named group updates
Message-id: <20191126133941.32492-55-mpetlan@redhat.com>
Patchwork-id: 289082
O-Subject: [RHEL8.2 BZ1743210 v2 054/327] sysfs: Fix internal_create_group() for named group updates
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit c855cf2759d27142f771173d9fd8e7fdf9cf5138
Author: Rajat Jain <rajatja@google.com>
Date:   Sat Jun 16 10:49:46 2018 -0700

description
===========
There are a couple of problems with named group updates in the code
today:

* sysfs_update_group() will always fail for a named group, because
  internal_create_group() will try to create a new sysfs directory
  unconditionally, which will ofcourse fail with -EEXIST.

* We can leak the kernfs_node for grp->name if some one tries to:
  - rename a group (change grp->name), or
  - update a named group, to an unnamed group

It appears that the whole purpose of sysfs_update_group() was to
allow changing the permissions or visibility of attributes and not
the names. So make it clear in the comments, and allow it to update
an existing named group.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>