[tools] perf intel-pt: Accumulate cycle count from CYC packets
Message-id: <20191126133941.32492-106-mpetlan@redhat.com>
Patchwork-id: 289107
O-Subject: [RHEL8.2 BZ1743210 v2 105/327] perf intel-pt: Accumulate cycle count from CYC packets
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 7b4b4f83881e11b1fe5d8743953f81addb0871de
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon May 20 14:37:11 2019 +0300

description
===========
In preparation for providing instructions-per-cycle (IPC) information,
accumulate cycle count from CYC packets.

Although CYC packets are optional (requires config term 'cyc' to enable
cycle-accurate mode when recording), the simplest way to count cycles is
with CYC packets.

The first complication is that cycles must be counted only when also
counting instructions.

That means when control flow packet generation is enabled i.e. between
TIP.PGE and TIP.PGD packets.

Also, sampling the cycle count follows the same rules as sampling the
timestamp, that is, not before the instruction to which the decoder is
walking is reached.

In addition, the cycle count is not accurate for any but the first
branch of a TNT packet.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>