[tools] perf augmented_raw_syscalls: Tell which args are filenames and how many bytes to copy
Message-id: <20191126133941.32492-124-mpetlan@redhat.com>
Patchwork-id: 289138
O-Subject: [RHEL8.2 BZ1743210 v2 123/327] perf augmented_raw_syscalls: Tell which args are filenames and how many bytes to copy
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 4cae8675ea798b141ccdeea0a5b3f46a1e4605eb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Jun 3 16:52:46 2019 -0300

description
===========
Since we know what args are strings from reading the syscall
descriptions in tracefs and also already mark such args to be beautified
using the syscall_arg__scnprintf_filename() helper, all we need is to
fill in this info in the 'syscalls' BPF map we were using to state which
syscalls the user is interested in, i.e. the syscall filter.

Right now just set that with PATH_MAX and unroll the syscall arg in the
BPF program, as the verifier isn't liking something clang generates when
unrolling the loop.

This also makes the augmented_raw_syscalls.c program support all arches,
since we removed that set of defines with the hard coded syscall
numbers, all should be automatically set for all arches, with the
syscall id mapping done correcly.

Doing baby steps here, i.e. just the first string arg for a syscall is
printed, syscalls with more than one, say, the various rename* syscalls,
need further work, but lets get first something that the BPF verifier
accepts before increasing the complexity

To test it, something like:

 # perf trace -e string -e /home/acme/git/perf/tools/perf/examples/bpf/augmented_raw_syscalls.c

With:

  # cat ~/.perfconfig
  [llvm]
	dump-obj = true
	clang-opt = -g
  [trace]
	#add_events = /home/acme/git/perf/tools/perf/examples/bpf/augmented_raw_syscalls.c
	show_zeros = yes
	show_duration = no
	no_inherit = yes
	show_timestamp = no
	show_arg_names = no
	args_alignment = 40
	show_prefix = yes
  #

That commented add_events line is needed for developing this
augmented_raw_syscalls.c BPF program, as if we add it via the
'add_events' mechanism so as to shorten the 'perf trace' command lines,
then we end up not setting up the -v option which precludes us having
access to the bpf verifier log :-\

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>