[tools] perf augmented_raw_syscalls: Change helper to consider just the augmented_filename part
Message-id: <20191126133941.32492-126-mpetlan@redhat.com>
Patchwork-id: 289139
O-Subject: [RHEL8.2 BZ1743210 v2 125/327] perf augmented_raw_syscalls: Change helper to consider just the augmented_filename part
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit deaf4da48a67f73eb7d5d1802c14eaf58d33f2da
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jun 4 15:57:28 2019 -0300

description
===========
So that we can use it for multiple args, baby steps not to step into the
verifier toes.

In the process make sure we handle -EFAULT from bpf_prog_read_str(), as
this really is needed now that we'll handle more than one augmented
argument, i.e. if there is failure, then we have the argument that fails
have:

  (size = 0, err = -EFAULT, value = [] )

followed by the next, lets say that worked for a second pathname:

  (size = 4, err = 0, value = "/tmp" )

So we can skip the first while telling the user about the problem and
then process the second.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>