[tools] perf cs-etm: Properly set the value of 'old' and 'head' in snapshot mode
Message-id: <20191126133941.32492-140-mpetlan@redhat.com>
Patchwork-id: 289188
O-Subject: [RHEL8.2 BZ1743210 v2 139/327] perf cs-etm: Properly set the value of 'old' and 'head' in snapshot mode
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit e45c48a9a4d20ebc7b639a62c3ef8f4b08007027
Author: Mathieu Poirier <mathieu.poirier@linaro.org>
Date: Wed Jun 5 10:16:33 2019 -0600

description
===========
This patch adds the necessary intelligence to properly compute the value
of 'old' and 'head' when operating in snapshot mode.  That way we can
get the latest information in the AUX buffer and be compatible with the
generic AUX ring buffer mechanic.

Tester notes:

> Leo, have you had the chance to test/review this one? Suzuki?

Sure.  I applied this patch on the perf/core branch (with latest
commit 3e4fbf36c1e3 'perf augmented_raw_syscalls: Move reading
filename to the loop') and passed testing with below steps:

  # perf record -e cs_etm/@tmc_etr0/ -S -m,64 --per-thread ./sort &
  [1] 19097
  Bubble sorting array of 30000 elements

  # kill -USR2 19097
  # kill -USR2 19097
  # kill -USR2 19097
  [ perf record: Woken up 4 times to write data ]
  [ perf record: Captured and wrote 0.753 MB perf.data ]

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>