[tools] perf intel-pt: Add reposition parameter to intel_pt_get_data()
Message-id: <20191126133941.32492-148-mpetlan@redhat.com>
Patchwork-id: 289205
O-Subject: [RHEL8.2 BZ1743210 v2 147/327] perf intel-pt: Add reposition parameter to intel_pt_get_data()
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 6c1f0b18ac3361837dbe53e794e28096285fb4f0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Jun 4 16:00:05 2019 +0300

description
===========
When the decoder gets the next trace buffer, some state is reset if the
buffer is not consecutive to the previous buffer. Add a parameter
'reposition' so that can be done also to support a "fast forward"
facility.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>