[tools] perf trace: Skip unknown syscalls when expanding strace like syscall groups
Message-id: <20191126133941.32492-164-mpetlan@redhat.com>
Patchwork-id: 289213
O-Subject: [RHEL8.2 BZ1743210 v2 163/327] perf trace: Skip unknown syscalls when expanding strace like syscall groups
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 04c41bcb862bbec1fb225243ecf07a3219593f81
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Jun 10 15:37:45 2019 -0300

description
===========
We have $INSTALL_DIR/share/perf-core/strace/groups/string files with
syscalls that should be selected when 'string' is used, meaning, in this
case, syscalls that receive as one of its arguments a string, like a
pathname.

But those were first selected and tested on x86_64, and end up failing
in architectures where some of those syscalls are not available, like
the 'access' syscall on arm64, which makes using 'perf trace -e string'
in such archs to fail.

Since this the routine doing the validation is used only when reading
such files, do not fail when some syscall is not found in the
syscalltbl, instead just use pr_debug() to register that in case people
are suspicious of problems.

Now using 'perf trace -e string' should work on arm64, selecting only
the syscalls that have a string and are available on that architecture.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>