[tools] perf tools: Don't hardcode host include path for libslang
Message-id: <20191126133941.32492-184-mpetlan@redhat.com>
Patchwork-id: 289163
O-Subject: [RHEL8.2 BZ1743210 v2 183/327] perf tools: Don't hardcode host include path for libslang
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 1955c8cf5e26b1f70d674190ff9984dbfd531ee9
Author: Florian Fainelli <f.fainelli@gmail.com>
Date: Fri Jun 14 11:39:47 2019 -0700

description
===========
Hardcoding /usr/include/slang is fundamentally incompatible with cross
compilation and will lead to the inability for a cross-compiled
environment to properly detect whether slang is available or not.

If /usr/include/slang is necessary that is a distribution specific
knowledge that could be solved with either a standard pkg-config .pc
file (which slang has) or simply overriding CFLAGS accordingly, but the
default perf Makefile should be clean of all of that.

Fixes: ef7b93a11904 ("perf report: Librarize the annotation code and use it in the newt browser")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>