[tools] tools build: Add test to check if slang.h is in /usr/include/slang/
Message-id: <20191126133941.32492-185-mpetlan@redhat.com>
Patchwork-id: 289305
O-Subject: [RHEL8.2 BZ1743210 v2 184/327] tools build: Add test to check if slang.h is in /usr/include/slang/
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit cbefd24f0aee3a5d787a013f207f6fd31d3c76d2
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jun 18 17:43:35 2019 -0300

description
===========
A few odd old distros (rhel5, 6, yeah, lots of those out in use, in many
cases we want to use upstream perf on it) have the slang header files in
/usr/include/slang/, so add a test that will be performed only when
test-all.c (the one with the most common sane settings) fails, either
because we're in one of these odd distros with slang/slang.h or because
something else failed (say libelf is not present).

So for the common case nothing changes, no additional test is performed.

Next step is to check in perf the result of these tests.

Fixes: 1955c8cf5e26 ("perf tools: Don't hardcode host include path for libslang")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>