[tools] perf intel-pt: Decoder to output CBR changes immediately
Message-id: <20191126133941.32492-191-mpetlan@redhat.com>
Patchwork-id: 289166
O-Subject: [RHEL8.2 BZ1743210 v2 190/327] perf intel-pt: Decoder to output CBR changes immediately
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit abe5a1d3e4bee361bd3b21b8909c8421e46911d1
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Sat Jun 22 12:32:42 2019 +0300

description
===========
The core-to-bus ratio (CBR) provides the CPU frequency. With branches
enabled, the decoder was outputting CBR changes only when there was a
branch. That loses the correct time of the change if the trace is not in
context (e.g. not tracing kernel space). Change to output the CBR change
immediately.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>