[tools] tools lib: Move argv_{split, free} from tools/perf/util/
Message-id: <20191126133941.32492-223-mpetlan@redhat.com>
Patchwork-id: 289231
O-Subject: [RHEL8.2 BZ1743210 v2 222/327] tools lib: Move argv_{split, free} from tools/perf/util/
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 9c10548c42219e961279826c2763a0e32dc056b9
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Jun 26 15:27:58 2019 -0300

description
===========
This came from the kernel lib/argv_split.c, so move it to
tools/lib/argv_split.c, to get it closer to the kernel structure.

We need to audit the usage of argv_split() to figure out if it is really
necessary to do have one allocation per argv[] entry, looking at one of
its users I guess that is not the case and we probably are even leaking
those allocations by not using argv_free() judiciously, for later.

With this we further remove stuff from tools/perf/util/, reducing the
perf specific codebase and encouraging other tools/ code to use these
routines so as to keep the style and constructs used with the kernel.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>