[tools] perf stat: Don't merge events in the same PMU
Message-id: <20191126133941.32492-225-mpetlan@redhat.com>
Patchwork-id: 289317
O-Subject: [RHEL8.2 BZ1743210 v2 224/327] perf stat: Don't merge events in the same PMU
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 6c5f4e5cb35b4694dc035d91092d23f596ecd06a
Author: Andi Kleen <ak@linux.intel.com>
Date: Mon Jun 24 12:37:09 2019 -0700

description
===========
Event merging is mainly to collapse similar events in lots of different
duplicated PMUs.

It can break metric displaying. It's possible for two metrics to have
the same event, and when the two events happen in a row the second
wouldn't be displayed.  This would also not show the second metric.

To avoid this don't merge events in the same PMU. This makes sense, if
we have multiple events in the same PMU there is likely some reason for
it (e.g. using multiple groups) and we better not merge them.

While in theory it would be possible to construct metrics that have
events with the same name in different PMU no current metrics have this
problem.

This is the fix for perf stat -M UPI,IPC (needs also another bug fix to
completely work)

Fixes: 430daf2dc7af ("perf stat: Collapse identically named events")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>