[tools] perf db-export: Pass main_thread to db_export__thread()
Message-id: <20191126133941.32492-277-mpetlan@redhat.com>
Patchwork-id: 289273
O-Subject: [RHEL8.2 BZ1743210 v2 276/327] perf db-export: Pass main_thread to db_export__thread()
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit ed5c0a16feb9f1a4347f109d5e9607f6f38688a0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Jul 10 11:57:52 2019 +0300

description
===========
Calls to db_export__thread() already have main_thread so there is no
reason to get it again, instead pass it as a parameter. Note that one
difference in this approach is that the main thread is not created if it
does not exist. It is better if it is not created because:

   - If main_thread is being traced it will have been created already.

   - If it is not being traced, there will be no other information about
     it, and it will never get deleted because there will be no EXIT event.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>