[tools] perf probe: Set pev->nargs to zero after freeing pev->args entries
Message-id: <20191126133941.32492-306-mpetlan@redhat.com>
Patchwork-id: 289310
O-Subject: [RHEL8.2 BZ1743210 v2 305/327] perf probe: Set pev->nargs to zero after freeing pev->args entries
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit df8350ed56a26f502a9636f37faf699a12ee906e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Jul 18 11:22:58 2019 -0300

description
===========
So that, when perf_add_probe_events() fails, like in:

  # perf probe icmp_rcv:64 "type=icmph->type"
  Failed to find 'icmph' in this function.
    Error: Failed to add events.
  Segmentation fault (core dumped)
  #

We don't segfault.

clear_perf_probe_event() was zeroing the whole pev, and since the switch
to zfree() for the members in the pev, that memset() was removed, which
left nargs with its original value, in the above case 1.

With the memset the same pev could be passed to clear_perf_probe_event()
multiple times, since all it would have would be zeroes, and free()
accepts zero, the loop would not happen and we would just memset it
again to zeroes.

Without it we got that segfault, so zero nargs to keep it like it was,
next cset will avoid calling clear_perf_probe_event() for the same pevs
in case of failure.

Fixes: d8f9da240495 ("perf tools: Use zfree() where applicable")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>