[tools] perf annotate: Fix printing of unaugmented disassembled instructions from BPF
Message-id: <20191126133941.32492-314-mpetlan@redhat.com>
Patchwork-id: 289337
O-Subject: [RHEL8.2 BZ1743210 v2 313/327] perf annotate: Fix printing of unaugmented disassembled instructions from BPF
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit 85127775a65fc58e69af0c44513937d471ccbe7b
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Aug 6 11:24:09 2019 -0300

description
===========
The code to disassemble BPF programs uses binutil's disassembling
routines, and those use in turn fprintf to print to a memstream FILE,
adding a newline at the end of each line, which ends up confusing the
TUI routines called from:

  annotate_browser__write()
    annotate_line__write()
      annotate_browser__printf()
        ui_browser__vprintf()
          SLsmg_vprintf()

The SLsmg_vprintf() function in the slang library gets confused with the
terminating newline, so make the disasm_line__parse() function that
parses the lines produced by the BPF specific disassembler (that uses
binutil's libopcodes) and the lines produced by the objdump based
disassembler used for everything else (and that doesn't adds this
terminating newline) trim the end of the line in addition of the
beginning.

This way when disasm_line->ops.raw, i.e. for instructions without a
special scnprintf() method, we'll not have that \n getting in the way of
filling the screen right after the instruction with spaces to avoid
leaving what was on the screen before and thus garbling the annotation
screen, breaking scrolling, etc.

Fixes: 6987561c9e86 ("perf annotate: Enable annotation of BPF programs")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>