[fs] ext4: enforce the immutable flag on open files
Message-id: <20191128114249.11162-11-lczerner@redhat.com>
Patchwork-id: 290508
O-Subject: [RHEL8.2 fs PATCH 10/18] ext4: enforce the immutable flag on open files
Bugzilla: 1775166
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Theodore Ts'o <tytso@mit.edu>

BZ 1775166
https://bugzilla.redhat.com/show_bug.cgi?id=1775166

BREW 24963453
https://brewweb.devel.redhat.com/taskinfo?taskID=24963453

commit 02b016ca7f99229ae6227e7b2fc950c4e140d74a

According to the chattr man page, "a file with the 'i' attribute
cannot be modified..."  Historically, this was only enforced when the
file was opened, per the rest of the description, "... and the file
can not be opened in write mode".

There is general agreement that we should standardize all file systems
to prevent modifications even for files that were opened at the time
the immutable flag is set.  Eventually, a change to enforce this at
the VFS layer should be landing in mainline.  Until then, enforce this
at the ext4 level to prevent xfstests generic/553 from failing.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: "Darrick J. Wong" <darrick.wong@oracle.com>
Cc: stable@kernel.org
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>