[powerpc] xive: Prevent page fault issues in the machine crash handler
Message-id: <20191128201337.31514-1-didoming@redhat.com>
Patchwork-id: 290534
O-Subject: [RHEL8.2 PATCH BZ1756116] powerpc/xive: Prevent page fault issues in the machine crash handler
Bugzilla: 1756116
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1756116
Brew:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=25017067
Tested: issues not observed in kernel 4.18.0-144.el8

When the machine crash handler is invoked, all interrupts are masked
but interrupts which have not been started yet do not have an ESB page
mapped in the Linux address space. This crashes the 'crash kexec'
sequence on sPAPR guests.

To fix, force the mapping of the ESB page when an interrupt is being
mapped in the Linux IRQ number space. This is done by setting the
initial state of the interrupt to OFF which is not necessarily the
case on PowerNV.

commit 1ca3dec2b2dff9d286ce6cd64108bda0e98f9710

Fixes: 243e25112d06 ("powerpc/xive: Native exploitation of the XIVE interrupt controller")
Cc: stable@vger.kernel.org # v4.12+
Signed-off-by: Cedric Le Goater <clg@kaod.org>
Reviewed-by: Greg Kurz <groug@kaod.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20191031063100.3864-1-clg@kaod.org
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>