[kernel] perf/aux: Fix AUX output stopping
Message-id: <20191205145753.18111-3-mpetlan@redhat.com>
Patchwork-id: 291191
O-Subject: [RHEL8.2 BZ1743210 v2 330/327] perf/aux: Fix AUX output stopping
Bugzilla: 1743210
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743210
https://bugzilla.redhat.com/show_bug.cgi?id=1743210

upstream
========
commit f3a519e4add93b7b31a6616f0b09635ff2e6a159
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date:   Tue Oct 22 10:39:40 2019 +0300

description
===========
Commit:

  8a58ddae2379 ("perf/core: Fix exclusive events' grouping")

allows CAP_EXCLUSIVE events to be grouped with other events. Since all
of those also happen to be AUX events (which is not the case the other
way around, because arch/s390), this changes the rules for stopping the
output: the AUX event may not be on its PMU's context any more, if it's
grouped with a HW event, in which case it will be on that HW event's
context instead. If that's the case, munmap() of the AUX buffer can't
find and stop the AUX event, potentially leaving the last reference with
the atomic context, which will then end up freeing the AUX buffer. This
will then trip warnings:

Fix this by using the context's PMU context when looking for events
to stop, instead of the event's PMU context.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>