[x86] kprobes: Verify stack frame on kretprobe
Message-id: <20191209203857.1447-2-jolsa@redhat.com>
Patchwork-id: 291505
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.2 BZ1779235 1/2] x86/kprobes: Verify stack frame on kretprobe
Bugzilla: 1779235
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

Bugzilla: 1779235
https://bugzilla.redhat.com/show_bug.cgi?id=1779235

upstream
========
commit 3ff9c075cc767b3060bdac12da72fc94dd7da1b8
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date:   Sun Feb 24 01:49:52 2019 +0900

description
===========
Verify the stack frame pointer on kretprobe trampoline handler,
If the stack frame pointer does not match, it skips the wrong
entry and tries to find correct one.

This can happen if user puts the kretprobe on the function
which can be used in the path of ftrace user-function call.
Such functions should not be probed, so this adds a warning
message that reports which function should be blacklisted.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>