[tools] perf probe: Show correct statement line number by perf probe -l
Message-id: <20191218212502.14325-16-mpetlan@redhat.com>
Patchwork-id: 292122
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 BZ1760227 v2 15/22] perf probe: Show correct statement line number by perf probe -l
Bugzilla: 1760227
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1760227
https://bugzilla.redhat.com/show_bug.cgi?id=1760227

upstream
========
commit 57f95bf5f88295612871c36cb0de5069e50570f8
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Mon Nov 18 17:11:50 2019 +0900

description
===========
The dwarf_getsrc_die() can return the line which is not a statement nor
the least line number among the lines which shares same address.

This can lead perf probe --list shows incorrect line number for probed
address.

To fix this, this introduces cu_getsrc_die() which returns only a
statement line and which is the least line number (we call it the
representive line for an address), and use it in cu_find_lineinfo().

Also, if the given address is the entry address of a real function,
cu_find_lineinfo() returns the function declared line number instead of
the start line number of the function body.

For example, without this change perf probe -l shows incorrect line as
below.

  # perf probe -a kernel_read:2
  Added new event:
    probe:kernel_read    (on kernel_read:2)

  You can now use it in all perf tools, such as:

  	perf record -e probe:kernel_read -aR sleep 1

  # perf probe -l
    probe:kernel_read    (on kernel_read:1@linux-5.0.0/fs/read_write.c)

With this fix, it shows correct line number as below;

  # perf probe -l
    probe:kernel_read    (on kernel_read:2@linux-5.0.0/fs/read_write.c)

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>