[tools] perf probe: Verify given line is a representive line
Message-id: <20191218212502.14325-17-mpetlan@redhat.com>
Patchwork-id: 292117
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 BZ1760227 v2 16/22] perf probe: Verify given line is a representive line
Bugzilla: 1760227
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1760227
https://bugzilla.redhat.com/show_bug.cgi?id=1760227

upstream
========
commit 1ae5d88a4eefacd4a3643170c20cf6403a24d254
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Mon Nov 18 17:12:00 2019 +0900

description
===========
Verify user given probe line is a representive line (which doesn't share
the address with other lines or the line is the least line among the
lines which shares same address), and if not, it shows what is the
representive line.

Without this fix, user can put a probe on the lines which is not a a
representive line. But since this is not a representive line, perf probe
-l shows a representive line number instead of user given line number.
e.g. (put kernel_read:3, but listed as kernel_read:2)

  # perf probe -a kernel_read:3
  Added new event:
    probe:kernel_read    (on kernel_read:3)

  You can now use it in all perf tools, such as:

  	perf record -e probe:kernel_read -aR sleep 1

  # perf probe -l
    probe:kernel_read    (on kernel_read:2@linux-5.0.0/fs/read_write.c)

With this fix, perf probe doesn't allow user to put a probe on a
representive line, and tell what is the representive line.

  # perf probe -a kernel_read:3
  This line is sharing the addrees with other lines.
  Please try to probe at kernel_read:2 instead.
    Error: Failed to add events.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>