[kernel] kprobes: Show address of kprobes if kallsyms does
Message-id: <20200109210415.10943-1-mpetlan@redhat.com>
Patchwork-id: 292319
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 PATCH] kprobes: Show address of kprobes if kallsyms does
Bugzilla: 1747099
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1747099
https://bugzilla.redhat.com/show_bug.cgi?id=1747099

upstream
========
commit 81365a947de453bcaba2558f8de5dadcffe05bc1
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date:   Sat Apr 28 21:36:02 2018 +0900

description
===========
Show probed address in debugfs kprobe list file as same
as kallsyms does. This information is used for checking
kprobes are placed in the expected address. So it should
be able to compared with address in kallsyms.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>