[redhat] kabi: Adding symbol __stack_chk_fail (kernel/panic.c)
Message-id: <20200122171848.9121-50-ckalina@redhat.com>
Patchwork-id: 293215
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 kABI 049/496] kabi: Adding symbol __stack_chk_fail (kernel/panic.c)
Bugzilla: 1794078
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Al Stone <ahs3@redhat.com>

This symbol is already whitelisted for non-aarch64 architectures.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

__attribute__ ( ( externally_visible ) ) void __stack_chk_fail ( void )

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>