[redhat] kabi: Adding symbol cpufreq_quick_get (drivers/cpufreq/cpufreq.c)
Message-id: <20200122171848.9121-129-ckalina@redhat.com>
Patchwork-id: 293445
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 kABI 128/496] kabi: Adding symbol cpufreq_quick_get (drivers/cpufreq/cpufreq.c)
Bugzilla: 1794078
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Al Stone <ahs3@redhat.com>

This symbol is already whitelisted for non-aarch64 architectures.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

__attribute__ ( ( always_inline , unused ) ) __attribute__ ( ( hotpatch
    ( 0 , 0 ) ) ) unsigned int cpufreq_quick_get ( unsigned int )

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>