[redhat] kabi: Adding symbol __ipv6_addr_type (net/ipv6/addrconf_core.c)
Message-id: <20200124103540.16303-4-ckalina@redhat.com>
Patchwork-id: 294034
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 kABI v2 03/30] kabi: Adding symbol __ipv6_addr_type (net/ipv6/addrconf_core.c)
Bugzilla: 1794073
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Partner Bugzilla: 1744222
https://bugzilla.redhat.com/show_bug.cgi?id=1744222

Description:
============
drbd_ko
In order to figure out whether or not two IPV6 addresses are equal, we
use the ipv6_addr_equal function. If they are equal, we look at the
address type of the two addresses, and if they're link local addresses,
we compare the scope ID. For that, we need to get the address type,
which is where ipv6_addr_type() comes in.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

int __ipv6_addr_type ( const struct in6_addr * )

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: Rashid Khan <rkhan@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>